// src/interfaces/modal-model.interface.ts
interface ModalModelInterface {
	/**
	 * Purpose:
	 *
	 * Interface for defining a login form,
	 * needed for typing/defining parameters of a login form.
	 */
	// button: {
	id: string;
	classList: Array<string>;
	title: string;
	body: string;
	actions: Array<string>;
	// };
}

export default ModalModelInterface;